requirements.txt
django>=1.9.8
sqlalchemy>=1.0.14
celery[redis]>=4.0
 djangorestframework==3.7.7
